% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ess.R
\name{ess}
\alias{ess}
\alias{ess,TwoStageBasket-method}
\title{Expected Sample Size}
\usage{
ess(design, ...)

\S4method{ess}{TwoStageBasket}(
  design,
  theta1 = NULL,
  n,
  n1,
  lambda,
  interim_fun,
  interim_params = list(),
  weight_fun,
  weight_params = list(),
  ...
)
}
\description{
Expected Sample Size
}
\section{Methods (by class)}{
\itemize{
\item \code{TwoStageBasket}: Expected sample size for two-stage basket design.
}}

