% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalweights.R
\name{globalweights_fix}
\alias{globalweights_fix}
\title{Fixed Global Weights}
\usage{
globalweights_fix(n, r, w)
}
\arguments{
\item{n}{The sample size per basket.}

\item{r}{Vector of responses.}

\item{w}{Fixed number with wich all weights are multiplied}
}
\value{
A numeric value.
}
\description{
Fixed Global Weights
}
\examples{
globalweights_fix(n = 20, r = c(1, 3, 5), w = 0.5)
}
