% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interim.R
\name{interim_postpred}
\alias{interim_postpred}
\alias{interim_postpred,TwoStageBasket-method}
\title{Interim Analysis Based on the Posterior Predictive Probability}
\usage{
interim_postpred(design, ...)

\S4method{interim_postpred}{TwoStageBasket}(
  design,
  n,
  n1,
  r1,
  lambda,
  weight_mat,
  prob_futstop = 0.1,
  prob_effstop = 0.9,
  ...
)
}
\description{
Interim Analysis Based on the Posterior Predictive Probability
}
\section{Methods (by class)}{
\itemize{
\item \code{TwoStageBasket}: Interim analysis based on the posterior
predictive probabilty for two-stage basket designs.
}}

