% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{setupOneStageBasket}
\alias{setupOneStageBasket}
\title{Setup OneStageBasket}
\usage{
setupOneStageBasket(k, shape1 = 1, shape2 = 1, theta0)
}
\arguments{
\item{k}{The number of baskets.}

\item{shape1}{First common shape parameter of the beta prior.}

\item{shape2}{Second common shape parameter of the beta prior.}

\item{theta0}{A common probability under the null hypothesis.}
}
\value{
An S4 object of class \code{\link{OneStageBasket}}.
}
\description{
Creates an object of class \code{\link{OneStageBasket}}.
}
\details{
A \code{\link{OneStageBasket}} object contains the most important
design features of a basket trial. Currently only common prior distributions
and a common null hypothesis are supported.
}
\examples{
design <- setupOneStageBasket(k = 3, theta0 = 0.2)
}
