% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{weights_cpp}
\alias{weights_cpp}
\title{Weights Based on the Calibrated Power Prior}
\usage{
weights_cpp(design, ...)
}
\arguments{
\item{design}{An object of class \code{Basket} created by
\code{setupBasket}.}

\item{...}{Further arguments.}
}
\value{
A matrix including the weights of all possible pairwise outcomes.
}
\description{
Weights Based on the Calibrated Power Prior
}
\examples{
design <- setupOneStageBasket(k = 3, theta0 = 0.2)
toer(design, n = 15, lambda = 0.99, weight_fun = weights_cpp)
}
