% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{weights_custom}
\alias{weights_custom}
\alias{weights_custom,OneStageBasket-method}
\title{Custom Weights}
\usage{
weights_custom(design, ...)

\S4method{weights_custom}{OneStageBasket}(design, n, mat, ...)
}
\arguments{
\item{design}{An object of class \code{Basket} created by
\code{setupOneStageBasket} or \code{setupTwoStageBasket}.}

\item{...}{Further arguments.}

\item{n}{The sample size per basket.}

\item{mat}{A matrix.}
}
\value{
A matrix including the weights of all possible pairwise outcomes.
}
\description{
Custom Weights
}
\details{
\code{weights_custom} simply returns the matrix that is passed
to \code{mat}. The input of the matrix is used as follows: The value
of the i-th row in the j-th column determines how much information is
shared between two baskets where one basket has i-1 responses and
the other basket hast j-1 responses. 1 is subtracted because 0 is also
a possible number of responses.
}
\section{Methods (by class)}{
\itemize{
\item \code{weights_custom(OneStageBasket)}: Custom weights for a single-stage basket design.

}}
\examples{
design <- setupOneStageBasket(k = 3, p0 = 0.2)
mat <- matrix(0.5, nrow = 16, ncol = 16)
toer(design, n = 15, lambda = 0.99, weight_fun = weights_custom,
  weight_params = list(mat = mat))
}
