% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{weights_eqprob}
\alias{weights_eqprob}
\alias{weights_eqprob,OneStageBasket-method}
\alias{weights_prob,OneStageBasket-method}
\title{Weights Based on the Equivalence Probability Weight}
\usage{
weights_eqprob(design, ...)

\S4method{weights_eqprob}{OneStageBasket}(design, n, bound, ...)

\S4method{weights_prob}{OneStageBasket}(design, n, ...)
}
\arguments{
\item{design}{An object of class \code{Basket} created by
\code{setupBasket}.}

\item{...}{Further arguments.}

\item{n}{The sample size per basket.}

\item{bound}{Equivalence bound.}
}
\value{
A matrix including the weights of all possible pairwise outcomes.
}
\description{
Weights Based on the Equivalence Probability Weight
}
\section{Methods (by class)}{
\itemize{
\item \code{OneStageBasket}: Equivalence probability weights for a
single-stage basket design.

\item \code{OneStageBasket}: Probability weights for a single-stage basket
design.
}}

