% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/weights.R
\name{weights_fujikawa}
\alias{weights_fujikawa}
\alias{weights_fujikawa,OneStageBasket-method}
\alias{weights_fujikawa,TwoStageBasket-method}
\title{Weights Based on Fujikawa et al.'s Design}
\usage{
weights_fujikawa(design, ...)

\S4method{weights_fujikawa}{OneStageBasket}(
  design,
  n,
  lambda,
  epsilon = 1.25,
  tau = 0.5,
  logbase = 2,
  prune = FALSE,
  ...
)

\S4method{weights_fujikawa}{TwoStageBasket}(
  design,
  n,
  n1,
  lambda,
  epsilon = 1.25,
  tau = 0,
  logbase = 2,
  prune = FALSE,
  ...
)
}
\description{
Weights Based on Fujikawa et al.'s Design
}
\section{Methods (by class)}{
\itemize{
\item \code{OneStageBasket}: Fujikawa-weights for a single-stage basket
design.

\item \code{TwoStageBasket}: Fujikawa-weights for a two-stage basket design.
}}

