% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/monotonicity.R
\name{check_mon_within}
\alias{check_mon_within}
\alias{check_mon_within,Basket-method}
\title{Check Within-Trial Monotonicity}
\usage{
check_mon_within(
  design,
  n,
  lambda,
  epsilon,
  tau,
  logbase = 2,
  prune,
  details,
  ...
)

\S4method{check_mon_within}{Basket}(
  design,
  n,
  lambda,
  epsilon,
  tau,
  logbase = 2,
  prune,
  details,
  ...
)
}
\arguments{
\item{design}{An object of class \code{Basket} created by
\code{setupBasket}.}

\item{n}{The sample size per basket.}

\item{lambda}{The posterior probability threshold. See details for more
information.}

\item{epsilon}{A tuning parameter that determines the amount of borrowing.
See details for more information.}

\item{tau}{A tuning parameter that determines how similar the baskets
have to be that borrowing occurs. See details for more information.}

\item{logbase}{A tuning parameter that determines which logarithm base is
used to compute the Jensen-Shannon divergence. See details for more
information.}

\item{prune}{Whether the results should be pruned before the final
analysis. See details for more information.}

\item{details}{Whether the cases where the monotonicity condition is
violated should be returned, in case there are any.}

\item{...}{Further arguments.}
}
\value{
If \code{details = FALSE} then only a logical value is returned.
If \code{details = TRUE} then if there are any cases where the
within-trial monotonicity condition is violated, a list of these cases and
their results are returned.
}
\description{
Checks whether the within-trial monotonicity condition holds.
}
\details{
\code{check_mon_within} checks whether the within-trial
monotonicity condition holds. For a single-stage design with equal
prior distributions and equal sample sizes for each basket this condition
states that there are no cases where the null hypothesis of a basket is
rejected when there is at least one other basket with more observed
responses for which the null hypothesis cannot be rejected.
}
\section{Methods (by class)}{
\itemize{
\item \code{Basket}: Within-trial monotonicity condition for a
single stage design.
}}

\examples{
design <- setupBasket(k = 4, shape1 = 1, shape2 = 1, theta0 = 0.2,
  theta1 = c(0.2, 0.5, 0.5, 0.5))
check_mon_within(design = design, n = 24, lambda = 0.99, epsilon = 0.5,
  tau = 0, prune = FALSE, details = TRUE)
}
