% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interim.R
\name{interim_posterior}
\alias{interim_posterior}
\alias{interim_posterior,TwoStageBasket-method}
\title{Interim analysis based on the posterior probability}
\usage{
interim_posterior(design, ...)

\S4method{interim_posterior}{TwoStageBasket}(design, n1, r1, weight_mat, prob_futstop, prob_effstop, ...)
}
\description{
Interim analysis based on the posterior probability
}
\section{Methods (by class)}{
\itemize{
\item \code{TwoStageBasket}: Interim analysis based on the posterior
probabilty for two-stage basket designs.
}}

