% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{interim_postpred}
\alias{interim_postpred}
\title{Interim Analysis Based on the Posterior Predictive Probability}
\usage{
interim_postpred(design, ...)
}
\description{
Interim Analysis Based on the Posterior Predictive Probability
}
