% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\docType{class}
\name{OneStageBasket-class}
\alias{OneStageBasket-class}
\alias{OneStageBasket}
\title{Class OneStageBasket}
\description{
OneStageBasket is an S4 class. An object of this class conatains the most
important design features of a single-stage basket trial.
}
\details{
Currently only common prior distributions and a common null
hypothesis are supported.
}
\section{Slots}{

\describe{
\item{\code{k}}{The number of baskets.}

\item{\code{shape1}}{First common shape parameter of the beta prior.}

\item{\code{shape2}}{Second common shape parameter of the beta prior.}

\item{\code{theta0}}{A common probability under the null hypothesis.}

\item{\code{theta1}}{A probability under the alternative hypothesis for each basket.}
}}

