% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\docType{class}
\name{OneStageBasket-class}
\alias{OneStageBasket-class}
\alias{OneStageBasket}
\title{Class OneStageBasket}
\description{
OneStageBasket is an S4 class. An object of this class contains the most
important design features of a single-stage basket trial.
}
\details{
This class implements a single-stage basket trial based on the design
proposed by Fujikawa et al. In this design, at first separate posterior
distributions are calculated for each basket based on a beta-binomial model.
Information is then borrowed between baskets by calculating weights that
reflect the similarity between the basket and computing posterior
distributions for each basket where the parameters of the beta posterior are
calculated as a weighted sum of the individual posterior distributions.
If for a basket the posterior probability that p > p0 is
greater than lambda, then the null hypothesis is rejected.

Currently only common prior distributions and a common null
hypothesis are supported.
}
\section{Slots}{

\describe{
\item{\code{k}}{The number of baskets.}

\item{\code{shape1}}{First common shape parameter of the beta prior.}

\item{\code{shape2}}{Second common shape parameter of the beta prior.}

\item{\code{p0}}{A common probability under the null hypothesis.}
}}

\references{
Fujikawa, K., Teramukai, S., Yokota, I., & Daimon, T. (2020).
A Bayesian basket trial design that borrows information across strata based
on the similarity between the posterior distributions of the response
probability. Biometrical Journal, 62(2), 330-338.
}
