% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_weights}
\alias{plot_weights}
\title{Plot Weight Functions}
\usage{
plot_weights(n, r1, weight_fun, weight_params = list())
}
\arguments{
\item{n}{The sample size per basket.}

\item{r1}{Number of responses in one basket}

\item{weight_fun}{Which function should be used to calculate the pairwise
weights.}

\item{weight_params}{A list of tuning parameters specific to
\code{weight_fun}.}
}
\value{
A plot.
}
\description{
Plot Weight Functions
}
\examples{
plot_weights(n = 20, r1 = 10, weight_fun = weights_prob)
}
