% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{setupOneStageBasket}
\alias{setupOneStageBasket}
\title{Setup OneStageBasket}
\usage{
setupOneStageBasket(k, shape1 = 1, shape2 = 1, theta0, theta1)
}
\arguments{
\item{k}{The number of baskets.}

\item{shape1}{First common shape parameter of the beta prior.}

\item{shape2}{Second common shape parameter of the beta prior.}

\item{theta0}{A common probability under the null hypothesis.}

\item{theta1}{A probability under the alternative hypothesis for each basket.}
}
\value{
An S4 object of class \code{OneStageBasket}.
}
\description{
Creates an object of class \code{OneStageBasket}.
}
\details{
A \code{OneStageBasket} object contains the most important design
features of a basket trial. Currently only common prior distributions and a
common null hypothesis are supported.
}
\examples{
design <- setupOneStageBasket(k = 3, theta0 = 0.2, theta1 = c(0.2, 0.5, 0.5))
}
