% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/toer.R
\name{toer}
\alias{toer}
\alias{toer,OneStageBasket-method}
\title{Type 1 Error Rate}
\usage{
toer(
  design,
  n,
  lambda,
  epsilon,
  tau,
  logbase = 2,
  prune = FALSE,
  results = c("fwer", "group"),
  ...
)

\S4method{toer}{OneStageBasket}(
  design,
  n,
  lambda,
  epsilon,
  tau,
  logbase = 2,
  prune = FALSE,
  results = c("fwer", "group"),
  ...
)
}
\arguments{
\item{design}{An object of class \code{Basket} created by
\code{setupBasket}.}

\item{n}{The sample size per basket.}

\item{lambda}{The posterior probability threshold. See details for more
information.}

\item{epsilon}{A tuning parameter that determines the amount of borrowing.
See details for more information.}

\item{tau}{A tuning parameter that determines how similar the baskets
have to be that borrowing occurs. See details for more information.}

\item{logbase}{A tuning parameter that determines which logarithm base is
used to compute the Jensen-Shannon divergence. See details for more
information.}

\item{prune}{Whether the results should be pruned before the final
analysis. See details for more information.}

\item{results}{Whether only the family wise error rate (option \code{fwer})
or also the rejection probabilities per group (option \code{group}) should
be returned.}

\item{...}{Further arguments.}
}
\value{
If \code{results = "fwer"} then the family wise type 1 error rate is
returned as a numeric value. If \code{results = "group"} then a list with
the rejection probabilities per group and the family wise type 1 error rate
is returned. If all theta1 > theta0 then the family wise type 1 error rate
is calculated under the global null hypothesis. For baskets with theta1 =
theta0 the rejection probabilities corresponds to the type 1 error rate, for
baskets with theta1 > theta0 the rejection probabilities corresponds to the
power.
}
\description{
Computes the exact family wise type 1 error rate of a basket trial .
}
\details{
\code{toer} computes the exact family wise type 1 error rate and the
exact rejection probabilities per group. The family wise type 1 error rate
is the probability to reject at least one null hypothesis for a basket with
theta1 = theta0. If all theta1 > theta0 then the family wise type 1 error
rate under the global null hypothesis is computed. The rejection
probabilities correspond to the type 1 error rate for baskets with theta1 =
theta 0 and to the power for baskets with theta1 > theta 0.

Calculations are based on the design of Fujikawa et al.:
At first separate posterior distributions are calculated for each basket
based on a beta-binomial model. Information is then borrowed between baskets
by calculating weights that reflect the similarity between the basket and
computing posterior distributions for each basket where the parameters
of the beta posterior are calculated as a weighted sum of the individual
posterior distributions. The weight between two baskets i and j is found as
(1 - JSD(i, j))^epsilon where JSD(i, j) is the Jensen-Shannon divergence
between basket i and j. A small value of epsilon results in stronger
borrowing also across baskets with heterogenous results. If epsilon is
large then information is only borrowed between baskets with similar results.
If a weight is smaller than tau it is set to 0, which results in no
borrowing. If for a basket the posterior probability that \eqn{\theta} >
theta0 is greater than lambda, then the null hypothesis is rejected.
}
\section{Methods (by class)}{
\itemize{
\item \code{OneStageBasket}: Type 1 error rate for a single-stage design.
}}

\examples{
design <- setupOneStageBasket(k = 3, theta0 = 0.2, theta1 = c(0.2, 0.5, 0.5))
toer(design, n = 15, lambda = 0.99, epsilon = 2, tau = 0)
}
\references{
Fujikawa, K., Teramukai, S., Yokota, I., & Daimon, T. (2020).
A Bayesian basket trial design that borrows information across strata based
on the similarity between the posterior distributions of the response
probability. Biometrical Journal, 62(2), 330-338.

If \code{prune = TRUE} then the baskets with an observed number of baskets
smaller than the pooled critical value are not borrowed from. The
pooled critical value is the smallest integer c for which all null
hypotheses can be rejected if the number of responses is exactly c for
all baskets.
}
