% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{weights_cpp}
\alias{weights_cpp}
\alias{weights_cpp,OneStageBasket-method}
\alias{weights_cpp,TwoStageBasket-method}
\title{Weights Based on the Calibrated Power Prior}
\usage{
weights_cpp(design, ...)

\S4method{weights_cpp}{OneStageBasket}(design, n, a = 1, b = 1, ...)

\S4method{weights_cpp}{TwoStageBasket}(design, n, n1, a = 1, b = 1, ...)
}
\arguments{
\item{design}{An object of class \code{Basket} created by
\code{setupBasket}.}

\item{...}{Further arguments.}

\item{n}{The sample size per basket.}

\item{a}{first tuning parameter}

\item{b}{second tuning parameter}

\item{n1}{The sample size per basket for the interim analysis in case of a
two-stage design.}
}
\value{
A matrix including the weights of all possible pairwise outcomes.
}
\description{
Weights Based on the Calibrated Power Prior
}
\section{Methods (by class)}{
\itemize{
\item \code{weights_cpp(OneStageBasket)}: Calibrated power prior weights for a single-stage
basket design.

\item \code{weights_cpp(TwoStageBasket)}: Calibrated power prior weights for a two-stage
basket design.

}}
\examples{
design <- setupOneStageBasket(k = 3, theta0 = 0.2)
toer(design, n = 15, lambda = 0.99, weight_fun = weights_cpp)
}
