% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{weights_mml}
\alias{weights_mml}
\alias{weights_mml,OneStageBasket-method}
\alias{weights_mml,TwoStageBasket-method}
\title{Weights Based on Maximising the Marginal Likelihood}
\usage{
weights_mml(design, ...)

\S4method{weights_mml}{OneStageBasket}(design, n, ...)

\S4method{weights_mml}{TwoStageBasket}(design, n, n1, ...)
}
\arguments{
\item{design}{An object of class \code{Basket} created by
\code{setupBasket}.}

\item{...}{Further arguments.}

\item{n}{The sample size per basket.}

\item{n1}{The sample size per basket for the interim analysis in case of a
two-stage design.}

\item{lambda}{The posterior probability threshold. See details for more
information.}
}
\value{
A matrix including the weights of all possible pairwise outcomes.
}
\description{
Weights Based on Maximising the Marginal Likelihood
}
\section{Methods (by class)}{
\itemize{
\item \code{OneStageBasket}: Maximum marginal likelihood weights for a
single-stage basket design.

\item \code{TwoStageBasket}: Maximum marginal likelihood weights for a
two-stage basket design.
}}

\examples{
design <- setupOneStageBasket(k = 3, theta0 = 0.2)
toer(design, n = 15, lambda = 0.99, weight_fun = weights_mml)
}
